+++ s/tools/maptool.ts
@@ -1,10 +1,12 @@
 
+/** @deprecated use `Map2` instead */
 export type MapBase<K, V> = {
 	has(key: K): boolean
 	get(key: K): V | undefined
 	set(key: K, value: V): void
 }
 
+/** @deprecated use `Map2` instead */
 export function maptool<K, V>(map: MapBase<K, V>) {
 	return {
 		guarantee: (key: K, make: () => V) => (
@@ -13,6 +15,7 @@ export function maptool<K, V>(map: MapBase<K, V>) {
 	}
 }
 
+/** @deprecated use `Map2` instead */
 export function mapGuarantee<K, V>(
 		map: MapBase<K, V>,
 		key: K,